Add convenience functionality to run templates

+++ runs/run_template/render_jobs.py
@@ -2,6 +2,22 @@ import os
 import jinja2
 import yaml
 
+import argparse
+
+my_parser = argparse.ArgumentParser()
+
+my_parser.add_argument('--NoPrep', action='store_true', default = False)
+my_parser.add_argument('--OverwriteJobs', action='store_true', default = False)
+
+args = vars(my_parser.parse_args())
+
+#Print args for debugging state
+print('-------INPUT PARAMS----------')
+for p in args.keys():
+    print('%s : %s'%(p.upper(), args[p]))
+print('-----------------------------')
+print('-----------------------------')
+
 tilenames = [
     'DES0544-2249',
     'DES2122+0001',
@@ -17,34 +33,34 @@ tilenames = [
     'DES2125-0124',
     'DES2125+0126',
     'DES2125-0207',
-    'DES2125+0209']
-# 'DES2128+0001',
-# 'DES2128-0041',
-# 'DES2128+0043',
-# 'DES2128-0124',
-# 'DES2128+0126',
-# 'DES2128-0207',
-# 'DES2128+0209',
-# 'DES2131+0001',
-# 'DES2131-0041',
-# 'DES2131+0043',
-# 'DES2131-0124',
-# 'DES2131+0126',
-# 'DES2131-0207',
-# 'DES2131+0209',
-# 'DES2134+0001',
-# 'DES2134-0041',
-# 'DES2134+0043',
-# 'DES2134-0124',
-# 'DES2134+0126',
-# 'DES2134-0207',
-# 'DES2134+0209',
-# 'DES2137+0001',
-# 'DES2137-0041',
-# 'DES2137+0043',
-# 'DES2137-0124',
-# 'DES2137+0126',
-# 'DES2137+0209']
+    'DES2125+0209',
+    'DES2128+0001',
+    'DES2128-0041',
+    'DES2128+0043',
+    'DES2128-0124',
+    'DES2128+0126',
+    'DES2128-0207',
+    'DES2128+0209',
+    'DES2131+0001',
+    'DES2131-0041',
+    'DES2131+0043']
+#    'DES2131-0124',
+#    'DES2131+0126',
+#    'DES2131-0207',
+#    'DES2131+0209',
+#    'DES2134+0001',
+#    'DES2134-0041',
+#    'DES2134+0043',
+#    'DES2134-0124',
+#    'DES2134+0126',
+#    'DES2134-0207',
+#    'DES2134+0209',
+#    'DES2137+0001',
+#    'DES2137-0041',
+#    'DES2137+0043',
+#    'DES2137-0124',
+#    'DES2137+0126',
+#    'DES2137+0209']
 
 if __name__ == '__main__':
     #Automatically get folder name (assuming certain folder structure)
@@ -81,9 +97,18 @@ if __name__ == '__main__':
         tmp = jinja2.Template(fp.read())
 
     for i, tilename in enumerate(tilenames):
-        gal_seed = 0
-        os.system('python $RUN_DIR/run_sims.py prep --tilename="%s" --bands="riz" --output-desdata="$PREP_DIR/%s/outputs_%s_seed%d_gplus"'%(tilename, name, tilename, gal_seed))
-        os.system('cp -r $PREP_DIR/%s/outputs_%s_seed%d_gplus $PREP_DIR/%s/outputs_%s_seed%d_gminus'%(name, tilename, gal_seed, name, tilename, gal_seed))
+        gal_seed = i
+        
+        if os.path.isfile('job_%s_minus.sh' % tilename) & os.path.isfile('job_%s_plus.sh' % tilename) & (args['OverwriteJobs'] == False):
+            print("----------------------------")
+            print("%s ALREADY PROCESSED"% tilename)
+            print("SKIPPING PROCESSING")
+            print("----------------------------")
+            continue
+
+        if args['NoPrep'] == False:
+            os.system('python $RUN_DIR/run_sims.py prep --tilename="%s" --bands="riz" --output-desdata="$PREP_DIR/%s/outputs_%s_seed%d_gplus"'%(tilename, name, tilename, gal_seed))
+            os.system('cp -r $PREP_DIR/%s/outputs_%s_seed%d_gplus $PREP_DIR/%s/outputs_%s_seed%d_gminus'%(name, tilename, gal_seed, name, tilename, gal_seed))
     
         with open('job_%s_plus.sh' % tilename, 'w') as fp:
             fp.write(tmp.render(tilename=tilename, model_name = name, plus_or_minus = "plus", seed_galsim=gal_seed, seed_mcal=42))
